+++ src/backend/resolvers/mutations/businessObject/businessObject/index.js
@@ -1,0 +1,5 @@
+const createBusinessObject = require('./create')
+
+module.exports = {
+  createBusinessObject,
+}