+++ src/backend/resolvers/mutations/businessObject/field/index.js
@@ -1,0 +1,5 @@
+const createBusinessObjectField = require('./create')
+
+module.exports = {
+  createBusinessObjectField,
+}