+++ src/backend/resolvers/mutations/businessObject/index.js
@@ -1,0 +1,7 @@
+const businessObjectResolvers = require('./businessObject')
+const businessObjectFieldResolvers = require('./field')
+
+module.exports = {
+  ...businessObjectResolvers,
+  ...businessObjectFieldResolvers,
+}