+++ src/backend/typeDefs/mutations/businessObject/businessObject/create.js
@@ -1,0 +1,10 @@
+const { gql } = require('apollo-server-express')
+
+const createBusinessObjectTypeDefs = gql`
+  input CreateBusinessObjectInput {
+    name: String!
+    sourceCollection: String
+  }
+`
+
+module.exports = createBusinessObjectTypeDefs