+++ src/backend/typeDefs/mutations/businessObject/businessObject/index.js
@@ -1,0 +1,5 @@
+const createBusinessObjectTypeDefs = require('./create')
+
+module.exports = [
+  createBusinessObjectTypeDefs,
+]