+++ src/backend/typeDefs/mutations/businessObject/field/create.js
@@ -1,0 +1,15 @@
+const { gql } = require('apollo-server-express')
+
+const createBusinessObjectFieldTypeDefs = gql`
+  input CreateBusinessObjectFieldInput {
+    businessObjectId: ID!
+    field: BoFieldInput
+  }
+
+  input BoFieldInput {
+    key: String!
+    type: String!
+  }
+`
+
+module.exports = createBusinessObjectFieldTypeDefs