+++ src/backend/typeDefs/mutations/businessObject/field/index.js
@@ -1,0 +1,5 @@
+const createBusinessObjectFieldTypeDefs = require('./create')
+
+module.exports = [
+  createBusinessObjectFieldTypeDefs,
+]