+++ src/backend/typeDefs/mutations/businessObject/index.js
@@ -1,0 +1,7 @@
+const businessObjectTypeDefs = require('./businessObject')
+const businessObjectFieldTypeDefs = require('./field')
+
+module.exports = [
+  ...businessObjectTypeDefs,
+  ...businessObjectFieldTypeDefs,
+]