+++ src/backend/typeDefs/mutations/index.js
@@ -18,6 +18,8 @@ const email = require('./email')
 
 const payerProjects = require('./payerProject')
 
+const businessObject = require('./businessObject')
+
 const mutationType = gql`
   type Mutation {
     createUser(input: CreateUserInput!): CreateUserPayload
@@ -107,6 +109,9 @@ const mutationType = gql`
     updatePayerProjectPtps(input: UpdatePayerProjectPtpsInput!): JSON
     removePayerProjectPtps(input: RemovePayerProjectPtpsInput!): [ID]!
     transferPayerProjectPtps(input: TransferPayerProjectPtpsInput!): JSON
+
+    createBusinessObject(input: CreateBusinessObjectInput!): BusinessObject
+    createBusinessObjectField(input: CreateBusinessObjectFieldInput!): BoField
   }
 `
 
@@ -130,4 +135,6 @@ module.exports = [
   ...workbook,
 
   ...payerProjects,
+
+  ...businessObject,
 ]